Fixes warnings due to httr::content not using default charset
For content type application/json the spec says that charset can be
omitted and if given it must be UTF-8. Unlike many widely used APIs,
the Phenoscape KB doesn't provide the charset attribute for
`application/json` content, and `httr::content()`, arguably erroneously,
does not assume UTF-8 in that case as the charset determined by the spec.

See phenoscape/phenoscape-kb-services#58 for discussion thread.