Fixes filtering of metadata tables returned from RNeXML
Beginning with RNeXML 2.3.0, tables returned from RNeXML::get_metadata()
are simplified by default. We could have forced the old format, but if
we need to make a change, then we might as well adapt to the new way of
doing things.

As a consequence, we have to require RNeXML 2.3.0 or higher.