Adds subsumer matrix query and some similarity metrics
The semantic similarity metrics included here are all solely based on
graph structure. That is, by default they will use a subsumer matrix
that has values 0 and 1, meaning each edge between a term and a
subsumer is given equal weight. The metrics are Jaccard, Tanimoto
(which is equal to Jaccard for W[i] in {0, 1}), and Cosine.

Tanimoto and Cosine could be used with weights other than {0, 1} as well,
such as Information Content.