Adds the ability to drop OTUs and characters from nexml
In contrast to subsetting a character matrix obtained from a nexml
object, this will allow subsetting the nexml object itself. If the
goal is to write out the subset object, then this will be far
superior to creating a nexml object from scratch and adding a subset
matrix to it. In large part this is because for the OTUs, characters,
and states retained it will also retain all their metadata, whereas
a character matrix extracted from a nexml object will have lost most
of the metadata, in particular for states and characters.