Enables switching to POST for JSON-returning queries
This had already been enabled for CSV returning queries.

Note that most KB API methods returning JSON do not actually support
POST at this point. However, the expectation is that for those that do
not, it is hardly possible to create a query URL (for a GET request)
that is too long. The one triggering the issue (/term/labels, see #113)
does support POST.

Fixes #113.