Allows controlling ontotrace behavior if names fail to resolve
See #118 for some context. Before this change, failure to resolve any
taxon name or entity name resulted in an empty matrix being returned. With
this change, in strict behavior (the default) it results in an error. If
non-strict behavior is selected, the query continues to be executed with
those taxon and entity terms that could be resolved to IRI.

Note that resolution by partial (rather than exact) match results in a
warning, but is still considered successful.