Adds ability to obtain or calculate term frequencies
Note that not all corpora are supported in the KB API yet. Specifically,
support for the "gene_annotations" corpus is pending in the API
(see phenoscape/phenoscape-kb-services#146 and phenoscape/phenoscape-kb-services#189),
and at least presently there are some issues in the KB API for the "genes"
corpus (see phenoscape/phenoscape-kb-services#191 and
phenoscape/phenoscape-kb-services#192).