Improves success rate of auto-determining term type
Some of these improvements, such as those based on decoding post-composed
term IRIs, will fail to work in the future when such term IRIs may be
opaque. Hopefully at that point the KB API will be able to resolve them
more properly, so that decoding IRIs, which is a hack anyway, won't be
needed then.