Introduces parameter to allow forcing a GET request
See #145 for context. In short, even some single URIs (in particular,
apparently, for phenotype objects) are longer than the 2KB threshold
for triggering the switch to HTTP POST, yet many API methods only accept
GET, not POST. This both increases the threshold by 50% for JSON data
requests, and introduces a parameter that if set to TRUE will force the
use of GET regardless of query length.

Fixes #145.