Provides an in-memory cache initializable from file
The main motivation for now is to help speed up computing term weights
for some of the corpora, in particular taxon_annotations, but it is
designed not to be specific to that case in any way.

Be default the cache will behave as an in-memory cache, i.e., starting
as an empty cache when the R session is started. Optionally it can,
however, be initialized with a precomputed cache, which would normally be
obtained by serializing the cache after the calculations or queries meant
to be cached have been run.