Change creation of resultant dataframe in mutual_exclusivity function.
In the mutual_exclusivity function, the resulting dataframe was created by first creating a dataframe with a single row and then appending the subsequent rows to it. Now, we create an empty dataframe (as opposed to a dataframe with a single row) and then append the subsequent rows to it.

Other change is that explicit integer conversion while storing the mutual exclusivity type in the resultan matrix in the mutual_exclusivity function is removed.