Filters out label-less terms from flagged ontologies
Some ontologies, in particular apparently CARO, have labels for terms
in the database only inconsistently. This filters out subsumer terms
that are in those ontologies, and failed to obtain a label. What is
really being tested here are post-compositions, so this shouldn't affect
the subsumers that are post-comps (whether they fail to obtain a label
or not).

Fixes #161. See also phenoscape/phenoscape-kb-services#378.