Adds Owl class for use with get_ontotrace_data()
Updates get_ontotrace_data() to support passing Manchester OWL
expressions via the Owl class for the taxon and entity parameters.

Adds as.owl() function that return an Owl object that can be
passed as taxon and entity to get_ontotrace_data().

The as.owl() function receives a character vector containing
Manchester OWL expressions and adds the "owl" class if necessary.
The as.owl() function has a format parameter that currently
must be "manchester" and apples a "owlmn" class. The format
parameter is for future expansion of the owl class.

The as.owl() function has an optional usesLabels parameter.
When usesLabels=TRUE as.owl() receives a character containing a
Manchester OWL expression containing labels instead
of IRI.  The /term/resolve_label_expression API endpoint is used
to resolve the label-based class expression into an valid
Manchester OWL expression and adds the "owl"/"owlmn" classes.

Adds subsume parameter to the get_ontotrace_data() to turn off
subsumption when passing in a vector of labels or IRI to
taxon/entity. This parameter defaults to TRUE and does not
work when passing owl data.

Co-authored-by: Hilmar Lapp <hlapp@drycafe.net>